增加了Common Patch Generation模块的实现，增加了RAHT模块的实现

+++ src/common/statistic.cc
@@ -178,10 +178,10 @@ float common::ParallelICPStat_t::dev_score(int type, int idx) {
 	return std::sqrt(var);
 }
 
-double common::Deviation(const std::vector<size_t>& _src) {
-	double sum      = std::accumulate(_src.begin(), _src.end(), 0);
-	double mean     = sum / _src.size();
-	double variance = 0.0;
+float common::Deviation(const std::vector<int>& _src) {
+	float sum      = std::accumulate(_src.begin(), _src.end(), 0);
+	float mean     = sum / _src.size();
+	float variance = 0.0;
 	std::for_each(_src.begin(), _src.end(), [&](const size_t x) { variance += std::pow(x - mean, 2); });
 	variance /= _src.size();
 	return std::sqrt(variance);
